rollback db when failed

+++ service/routes.py
@@ -32,7 +32,7 @@ from . import status  # HTTP Status Codes
 # Import Flask application
 from service.models.wishlist import Wishlist, WishlistVo
 from service.models.product import Product
-from service.models.model_utils import Availability, InCartStatus, get_non_null_product_fields, db
+from service.models.model_utils import Availability, DataValidationError, InCartStatus, get_non_null_product_fields, db
 
 ######################################################################
 # GET INDEX
@@ -167,12 +167,15 @@ class WishlistResource(Resource):
   # RETRIEVE A Wishlist
   # ------------------------------------------------------------------
   @api.doc('get_wishlists')
+  @api.response(400, 'Integer value expected for field: Wishlist ID')
   @api.response(404, 'Wishlist not found')
   @api.marshal_with(wishlist_vo)
   def get(self, wishlist_id):
     """
     Lists all products in a wishlist based on a wishlist_id
     """
+    if not wishlist_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, "Integer value expected for field: Wishlist ID")
     app.logger.info("Request to list products in a wishlist")
     wishlist = Wishlist.find_by_id(wishlist_id)
     if not wishlist:
@@ -201,6 +204,9 @@ class WishlistResource(Resource):
         status.HTTP_415_UNSUPPORTED_MEDIA_TYPE, \
         "Unsupported media type : application/json expected"
       )
+
+    if not wishlist_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, "Integer value expected for field: Wishlist ID")
 
     wishlist = Wishlist.find_by_id(wishlist_id)
     if not wishlist:
@@ -244,12 +250,16 @@ class WishlistResource(Resource):
   #------------------------------------------------------------------
   @api.doc('delete_wishlists')
   @api.response(204, 'Wishlist deleted')
+  @api.response(400, 'Integer value expected for field: Wishlist ID')
   def delete(self, wishlist_id):
     """
     Deletes a wishlist
     This endpoint will delete a wishlist based the id specified in the URL
     """
     app.logger.info("Request to delete wishlist with id: %s", wishlist_id)
+    if not wishlist_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, "Integer value expected for field: Wishlist ID")
+
     wishlist = Wishlist.find_by_id(wishlist_id)
     if wishlist:
       wishlist.delete()
@@ -321,7 +331,7 @@ class WishlistCollection(Resource):
     app.logger.info("Request to create a wishlist")
     # Check for form submission data
     if request.headers.get("Content-Type") != "application/json":
-      return abort(
+      abort(
         status.HTTP_415_UNSUPPORTED_MEDIA_TYPE, \
         "Unsupported media type : application/json expected"
       )
@@ -365,7 +375,7 @@ class ProductCollectionResource(Resource):
     """
     app.logger.info("Request to create a product")
     if request.headers.get("Content-Type") != "application/json":
-      return abort(
+      abort(
         status.HTTP_415_UNSUPPORTED_MEDIA_TYPE, \
         "Unsupported media type : application/json expected"
       )
@@ -375,6 +385,9 @@ class ProductCollectionResource(Resource):
 
     if not isinstance(data, dict):
       abort(status.HTTP_400_BAD_REQUEST, "Expected a json request body")
+
+    if not wishlist_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, "Integer value expected for field: Wishlist ID")
 
     data['wishlist_id'] = wishlist_id
 
@@ -387,12 +400,16 @@ class ProductCollectionResource(Resource):
     return product.serialize(), status.HTTP_201_CREATED, {'Location':location_url}
 
   @api.doc('delete_all_products_from_a_wishlist')
-  @api.response(204,"Products deleted")
+  @api.response(204, "Products deleted")
+  @api.response(400, 'Integer value expected for field: Wishlist ID')
   def delete(self, wishlist_id):
     """
     Deletes all products from a wishlist
     This endpoint will delete all products in a wishlist
     """
+    if not wishlist_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, "Integer value expected for field: Wishlist ID")
+
     wishlist = Wishlist.find_by_id(wishlist_id)
 
     if wishlist:
@@ -421,6 +438,7 @@ class ProductResource(Resource):
   PUT - Update a product in a wishlist
   """
   @api.doc('return_one_product')
+  @api.response(400, 'Integer value expected for fields: Wishlist ID and Product ID')
   @api.response(404, 'Product with id not found in wishlist with id')
   @api.marshal_with(full_product_model)
   def get(self,wishlist_id,product_id):
@@ -430,6 +448,9 @@ class ProductResource(Resource):
     Then look for a product based on a product_id
     """
     app.logger.info("Request to get a specific product in a wishlist")
+    if not wishlist_id.isdigit() or not product_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, 'Integer field expected for fields: Wishlist ID and Product ID')
+
     wishlist_product = Product.find_by_wishlist_id_and_product_id(wishlist_id, product_id)
 
     if not wishlist_product:
@@ -446,12 +467,16 @@ class ProductResource(Resource):
 
   @api.doc('delete_a_product')
   @api.response(204, "Product deleted")
+  @api.response(400, 'Integer value expected for fields: Wishlist ID and Product ID')
   def delete(self, wishlist_id, product_id):
     """
     Deletes a product from a wishlist
     This endpoint will delete an existing product in a wishlist
     """
     app.logger.info(f"Request to delete products with id {product_id} from wishlist {wishlist_id}")
+    if not wishlist_id.isdigit() or not product_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, 'Integer field expected for fields: Wishlist ID and Product ID')
+
     wishlist = Wishlist.find_by_id(wishlist_id)
     if wishlist:
       wishlist.delete_products([product_id])
@@ -470,6 +495,9 @@ class ProductResource(Resource):
     This endpoint will modify an existing product in a wishlist
     """
     app.logger.info("Request to update a product")
+    if not wishlist_id.isdigit() or not product_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, 'Integer field expected for fields: Wishlist ID and Product ID')
+
     product = Product.find_by_wishlist_id_and_product_id(wishlist_id, product_id)
     if request.headers.get("Content-Type") != "application/json":
       abort(
@@ -504,6 +532,7 @@ class AddToCartResource(Resource):
   """
 
   @api.doc('place_a_product_to_shopping_cart')
+  @api.response(400, "Integer value expected for fields: Wishlist ID and Product ID")
   @api.response(404, "Product not found in the wishlist")
   @api.marshal_with(full_product_model)
   def put(self, wishlist_id, product_id):
@@ -512,6 +541,9 @@ class AddToCartResource(Resource):
     This endpoint will move an existing product in a wishlist to a shopping cart
     """
     app.logger.info("Request to place a product in wishlist to cart")
+    if not wishlist_id.isdigit() or not product_id.isdigit():
+      abort(status.HTTP_400_BAD_REQUEST, 'Integer field expected for fields: Wishlist ID and Product ID')
+
     product = Product.find_by_wishlist_id_and_product_id(wishlist_id, product_id)
 
     if not product:
@@ -521,6 +553,15 @@ class AddToCartResource(Resource):
     product.update()
 
     return product.serialize(),status.HTTP_200_OK
+
+
+@api.errorhandler(DataValidationError)
+def handle_data_validation_error(error):
+  return {"message":error.args[0]}, status.HTTP_400_BAD_REQUEST
+
+@api.errorhandler(TypeError)
+def handle_type_error(error):
+  return {"message":error.args[0]}, status.HTTP_400_BAD_REQUEST
 
 ######################################################################
 # INITIALIZE DATABASE
